requirements.txt
Django==1.10.5
djangorestframework==3.5.4
matplotlib==2.0.0
numpy==1.12.0
pandas==0.19.2
Pillow==4.0.0
requests==2.13.0
celery==3.1.25
django-celery==3.2.1